feat(network): add generic event common to behaviour and handler (#1207)
