+++ crates/papyrus_network/src/streamed_data_protocol/mod.rs
@@ -15,6 +15,14 @@ pub struct InboundSessionId {
     value: usize,
 }
 
+#[derive(Debug)]
+// TODO(shahak) remove allow(dead_code).
+#[allow(dead_code)]
+pub(crate) enum SessionId {
+    OutboundSessionId(OutboundSessionId),
+    InboundSessionId(InboundSessionId),
+}
+
 // This is a workaround for the unstable feature trait aliases
 // https://doc.rust-lang.org/beta/unstable-book/language-features/trait-alias.html
 pub(crate) trait QueryBound: Message + 'static + Default {}
@@ -22,3 +30,12 @@ impl<T> QueryBound for T where T: Message + 'static + Default {}
 
 pub(crate) trait DataBound: Message + 'static + Unpin {}
 impl<T> DataBound for T where T: Message + 'static + Unpin {}
+
+#[derive(Debug)]
+// TODO(shahak) remove allow dead code.
+#[allow(dead_code)]
+pub(crate) enum GenericEvent<Query: QueryBound, Data: DataBound, SessionError> {
+    NewInboundSession { query: Query, inbound_session_id: InboundSessionId },
+    ReceivedData { outbound_session_id: OutboundSessionId, data: Data },
+    SessionFailed { session_id: SessionId, error: SessionError },
+}
